requirements.txt
alembic==1.16.2
atpublic==6.0.1
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Admin==2.0.2
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
